+++ NEWS.md
@@ -1,5 +1,10 @@
 # CytoPipelineUtils 0.99
 
-## CytoPipeline 0.99.0
+## CytoPipelineUtils 0.99.1
+
+- added support for FlowJo manual gating as a CytiProcessingStep
+- used file.path() instead of paste0() to construct file paths 
+
+## CytoPipelineUtils 0.99.0
 
 - First version